requirements.txt
Flask==2.3.3
gunicorn==21.2.0
pyswisseph==2.10.3.2
pytz==2023.3
supabase==2.0.2
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.10.0
timezonefinder==6.2.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
jinja2==3.1.2
redis==5.0.1